+++ modules/sde-core/src/main/java/org/eclipse/tractusx/sde/core/failurelog/repository/ConsumerDownloadHistoryRepository.java
@@ -1,6 +1,6 @@
 /********************************************************************************
- * Copyright (c) 2022, 2023 T-Systems International GmbH
- * Copyright (c) 2022, 2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2024 T-Systems International GmbH
+ * Copyright (c) 2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -18,27 +18,14 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-package org.eclipse.tractusx.sde.core.processreport.entity;
-
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-
-import jakarta.persistence.AttributeConverter;
-import jakarta.persistence.Converter;
+package org.eclipse.tractusx.sde.core.failurelog.repository;
 
 
-@Converter
-public class ListToStringConverter implements AttributeConverter<List<String>, String> {
-	
-    @Override
-    public String convertToDatabaseColumn(List<String> attribute) {
-        return attribute == null ? null : String.join(",",attribute);
-    }
+import org.eclipse.tractusx.sde.core.processreport.entity.ConsumerDownloadHistoryEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-    @Override
-    public List<String> convertToEntityAttribute(String dbData) {
-        return dbData == null ? Collections.emptyList() : Arrays.asList(dbData.split(","));
-    }
-    
+public interface ConsumerDownloadHistoryRepository extends JpaRepository<ConsumerDownloadHistoryEntity, String> {
+
+	ConsumerDownloadHistoryEntity findByProcessId(String id);
+
 }