+++ modules/sde-submodules/serial-part-typization/src/main/java/org/eclipse/tractusx/sde/submodels/spt/steps/DigitalTwinsAspectCsvHandlerUseCase.java
@@ -70,7 +70,7 @@ public class DigitalTwinsAspectCsvHandlerUseCase extends Step {
 		String shellId;
 		
 		ShellDescriptorRequest aasDescriptorRequest = digitalTwinsUtility.getShellDescriptorRequest(
-				aspect.getNameAtManufacturer(), aspect.getManufacturerPartId(), aspect.getUuid(),
+				aspect.getNameAtManufacturer()+"_"+aspect.getPartInstanceId(), aspect.getManufacturerPartId(), aspect.getUuid(),
 				getSpecificAssetIds(aspect), policy);
 
 		if (shellIds.isEmpty()) {
@@ -100,8 +100,9 @@ public class DigitalTwinsAspectCsvHandlerUseCase extends Step {
 		if (subModelResponse == null || foundSubmodel == null) {
 			logDebug(String.format("No submodels for '%s'", shellId));
 			
-			CreateSubModelRequest createSubModelRequest = digitalTwinsUtility
-					.getCreateSubModelRequest(aspect.getShellId(), getsemanticIdOfModel(), getIdShortOfModel());
+			CreateSubModelRequest createSubModelRequest = digitalTwinsUtility.getCreateSubModelRequest(
+					aspect.getShellId(), getsemanticIdOfModel(), getIdShortOfModel(), getNameOfModel(),
+					aspect.getUuid());
 			
 			digitalTwinsFacilitator.updateShellDetails(shellId, aasDescriptorRequest, createSubModelRequest);