+++ modules/sde-submodules/single-level-usage-as-built/src/main/java/org/eclipse/tractusx/sde/submodels/sluab/steps/EDCSingleLevelUsageAsBuiltHandlerUseCase.java
@@ -24,11 +24,13 @@ import java.util.Map;
 import org.eclipse.tractusx.sde.common.constants.CommonConstants;
 import org.eclipse.tractusx.sde.common.entities.PolicyModel;
 import org.eclipse.tractusx.sde.common.exception.CsvHandlerUseCaseException;
+import org.eclipse.tractusx.sde.common.exception.NoDataFoundException;
 import org.eclipse.tractusx.sde.common.exception.ServiceException;
 import org.eclipse.tractusx.sde.common.submodel.executor.Step;
 import org.eclipse.tractusx.sde.edc.entities.request.asset.AssetEntryRequest;
 import org.eclipse.tractusx.sde.edc.entities.request.asset.AssetEntryRequestFactory;
 import org.eclipse.tractusx.sde.edc.facilitator.CreateEDCAssetFacilator;
+import org.eclipse.tractusx.sde.edc.facilitator.DeleteEDCFacilitator;
 import org.eclipse.tractusx.sde.edc.gateways.external.EDCGateway;
 import org.eclipse.tractusx.sde.submodels.sluab.entity.SingleLevelUsageAsBuiltEntity;
 import org.eclipse.tractusx.sde.submodels.sluab.model.SingleLevelUsageAsBuilt;
@@ -37,7 +39,9 @@ import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class EDCSingleLevelUsageAsBuiltHandlerUseCase extends Step {
@@ -46,6 +50,7 @@ public class EDCSingleLevelUsageAsBuiltHandlerUseCase extends Step {
 	private final EDCGateway edcGateway;
 	private final CreateEDCAssetFacilator createEDCAssetFacilator;
 	private final SingleLevelUsageAsBuiltService singleLevelUsageAsBuiltService;
+	private final DeleteEDCFacilitator deleteEDCFacilitator;
 
 	@SneakyThrows
 	public SingleLevelUsageAsBuilt run(String submodel, SingleLevelUsageAsBuilt input, String processId,
@@ -57,16 +62,22 @@ public class EDCSingleLevelUsageAsBuiltHandlerUseCase extends Step {
 
 			AssetEntryRequest assetEntryRequest = assetFactory.getAssetRequest(submodel,
 					getSubmodelShortDescriptionOfModel(), shellId, subModelId, input.getParentUuid());
+
+			Map<String, String> eDCAsset =null;
+			
 			if (!edcGateway.assetExistsLookup(assetEntryRequest.getId())) {
-
-				edcProcessingforAspectRelationship(assetEntryRequest, input, policy);
-
+				eDCAsset = createEDCAssetFacilator.createEDCAsset(assetEntryRequest, policy);
 			} else {
-
-				deleteEDCFirstForUpdate(submodel, input, processId);
-				edcProcessingforAspectRelationship(assetEntryRequest, input, policy);
+			    eDCAsset = createEDCAssetFacilator.updateEDCAsset(assetEntryRequest, policy);
 				input.setUpdated(CommonConstants.UPDATED_Y);
 			}
+			
+			// EDC transaction information for DB
+			input.setAssetId(eDCAsset.get("assetId"));
+			input.setAccessPolicyId(eDCAsset.get("accessPolicyId"));
+			input.setUsagePolicyId(eDCAsset.get("usagePolicyId"));
+			input.setContractDefinationId(eDCAsset.get("contractDefinitionId"));
+			
 
 			return input;
 		} catch (Exception e) {
@@ -80,6 +91,9 @@ public class EDCSingleLevelUsageAsBuiltHandlerUseCase extends Step {
 			SingleLevelUsageAsBuiltEntity aspectRelationshipEntity = singleLevelUsageAsBuiltService
 					.readEntity(input.getChildUuid());
 			singleLevelUsageAsBuiltService.deleteEDCAsset(aspectRelationshipEntity);
+		} catch (NoDataFoundException e) {
+			log.warn("The EDC assetInfo not found in local database for delete, looking EDC connector and going to delete asset");
+			deleteEDCFacilitator.findEDCOfferInformation(input.getShellId(), input.getSubModelId());
 		} catch (Exception e) {
 			if (!e.getMessage().contains("404 Not Found")) {
 				throw new ServiceException("Unable to delete EDC offer for update: " + e.getMessage());