+++ CHANGELOG.md
@@ -2,12 +2,16 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
-
 ## [Unreleased]
 
 ### Added
 - Updated readme.md and open api yml file.
-
+- Added controller interface api's for Policy management.
+- External EDC service interface api updated.
+- Updated supported sub-model implementation classes.
+- EDC asset update refactored in supported submodels.
+- Support for pcf v6.0.0 submodel.
+- Added controller interface api's for PCF Exchange.
 
 ## [2.3.6] - 2024-03-06
 ### Fixed